feature(cookies-policy): move cookies to a hook
